Move RecourceCard components to top level resources components folder
