+++ src/app/(main-layout)/resources/_components/ResourceCard.tsx
