+++ src/app/(main-layout)/resources/_components/ResourceVoteControls.tsx
