+++ src/app/(main-layout)/resources/_components/ResourcesSearchResults.tsx
@@ -3,7 +3,7 @@
 import Paginator from "@/components/Paginator";
 import Resource, { ResourceSearchParams } from "@/types/Resource";
 import { SearchData } from "@/types/Search";
-import ResourceCard from "../[id]/_components/ResourceCard";
+import ResourceCard from "./ResourceCard";
 import { domAnimation, LazyMotion, m } from "framer-motion";
 import { searchResources } from "@/lib/actions/resource";
