+++ src/app/(main-layout)/resources/_components/SocialMediaLink.tsx
